moved datetime type, updated context
